Added geotiff to datalakes outputs

+++ adapters/datalakes/datalakes.py
@@ -62,31 +62,31 @@ def apply(env, params, l2product_files, date):
             satellite = get_satellite_name_from_product_name(os.path.basename(l2product_file))
             date = get_sensing_datetime_from_product_name(os.path.basename(l2product_file))
             out_path = os.path.join(os.path.dirname(os.path.dirname(l2product_file)), OUT_DIR, "datalakes", params['General']['wkt_name'], satellite + "_" + date)
-            output_file_main = os.path.join(out_path, NC_FILENAME.format(processor, satellite, date))
+            input_file = os.path.join(out_path, NC_FILENAME.format(processor, satellite, date))
             os.makedirs(out_path, exist_ok=True)
             bands_list = list(filter(None, params[PARAMS_SECTION][key].split(",")))
             bands, bands_min, bands_max = parse_bands(bands_list)
 
-            if os.path.exists(output_file_main):
+            if os.path.exists(input_file):
                 if ("synchronise" in params["General"].keys() and params['General']['synchronise'] == "false") or \
                         ("synchronise" in params["DATALAKES"].keys() and params["DATALAKES"]["synchronise"] == "false"):
-                    log(env["General"]["log"], "Removing file: ${}".format(output_file_main), indent=2)
-                    os.remove(output_file_main)
+                    log(env["General"]["log"], "Removing file: ${}".format(input_file), indent=2)
+                    os.remove(input_file)
 
-            if os.path.exists(output_file_main):
+            if os.path.exists(input_file):
                 log(env["General"]["log"], "Skipping processor {}. Target already exists".format(processor), indent=2)
             else:
                 log(env["General"]["log"], "Copying {} to Datalakes folder.".format(os.path.basename(l2product_file)), indent=2)
                 with open(l2product_file, "rb") as f:
                     nc_bytes = f.read()
-                with open(output_file_main, "wb") as f:
+                with open(input_file, "wb") as f:
                     f.write(nc_bytes)
 
                 if "S3" in satellite:
                     try:
-                        log(env["General"]["log"], "Merging {} with lake_mask_sui_S3.nc".format(os.path.basename(output_file_main)), indent=2)
+                        log(env["General"]["log"], "Merging {} with lake_mask_sui_S3.nc".format(os.path.basename(input_file)), indent=2)
                         lake_mask = get_pixels_from_nc(os.path.join(os.path.abspath(os.path.dirname(__file__)), "lake_mask_sui_S3.nc"), "Swiss_S3_water")
-                        with Dataset(output_file_main, mode='r+') as dst:
+                        with Dataset(input_file, mode='r+') as dst:
                             create_band(dst, "lake_mask", "", "lake_mask>0")
                             write_all_pixels_to_nc(dst, "lake_mask", lake_mask)
                             append_to_valid_pixel_expression(dst, "lake_mask>0")
@@ -96,14 +96,19 @@ def apply(env, params, l2product_files, date):
                 for idx, val in enumerate(bands):
                     log(env["General"]["log"], "Converting {} band {} to JSON".format(processor, val), indent=3)
                     if "S3" in satellite:
-                        output_file = os.path.join(out_path, JSON_FILENAME.format(processor, val, satellite, date))
-                        convert_nc("json", output_file_main, output_file, val, 6, bands_min[idx], bands_max[idx], satellite, date, env)
+                        json_outfile = os.path.join(out_path, JSON_FILENAME.format(processor, val, satellite, date))
+                        geotiff_outfile = os.path.join(out_path, GEOTIFF_FILENAME.format(processor, val, satellite, date, "sui"))
+                        convert_nc("json", input_file, json_outfile, val, 6, bands_min[idx], bands_max[idx], satellite, date, env)
+                        convert_nc("geotiff", input_file, geotiff_outfile, val, 6, bands_min[idx], bands_max[idx], satellite, date, env)
                     if "S2" in satellite:
                         tile = l2product_file.split("_")[-2]
-                        output_file = os.path.join(out_path, GEOTIFF_FILENAME.format(processor, val, satellite, date, tile))
-                    else:
-                        output_file = os.path.join(out_path, GEOTIFF_FILENAME.format(processor, val, satellite, date, "sui"))
-                    convert_nc("geotiff", output_file_main, output_file, val, 6, bands_min[idx], bands_max[idx], satellite, date, env)
+                        geotiff_outfile = os.path.join(out_path, GEOTIFF_FILENAME.format(processor, val, satellite, date, tile))
+                        if processor == "ACOLITE":
+                            convert_nc("geotiff", input_file, geotiff_outfile, val, 6, bands_min[idx], bands_max[idx],
+                                       satellite, date, env, projection=32631)
+                        else:
+                            convert_nc("geotiff", input_file, geotiff_outfile, val, 6, bands_min[idx], bands_max[idx],
+                                       satellite, date, env)
 
             if "bucket" not in params[PARAMS_SECTION]:
                 raise ValueError("S3 Bucket must be defined in parameters file")
@@ -126,21 +131,18 @@ def apply(env, params, l2product_files, date):
                              env[PARAMS_SECTION]["aws_secret_access_key"], env["General"]["log"])
 
         log(env["General"]["log"], "Notifying Datalakes API of updated data.", indent=1)
-        # requests.get(NOTIFY_URL)
+        requests.get(NOTIFY_URL)
 
 
-def convert_nc(output_type, input_file, output_file, band, decimals, band_min, band_max, satellite, date, env):
+def convert_nc(output_type, input_file, output_file, band, decimals, band_min, band_max, satellite, date, env, projection=4326):
+    log(env["General"]["log"], "Reading data from {}".format(input_file), indent=4)
     valid_pixel_expression = ""
     with Dataset(input_file, "r", format="NETCDF4") as nc:
         variables = nc.variables.keys()
-        if "lat" in variables:
-            lat = np.array(nc.variables["lat"][:])
-            lon = np.array(nc.variables["lon"][:])
-        else:
-            lat = np.array(nc.variables["latitude"][:])
-            lon = np.array(nc.variables["longitude"][:])
         values = np.array(nc.variables[band][:])
         values_flat = values.flatten()
+        y = np.array(nc.variables[nc.variables[band].dimensions[0]][:])
+        x = np.array(nc.variables[nc.variables[band].dimensions[1]][:])
         try:
             valid_pixel_expression = nc.variables[band].valid_pixel_expression
             vpe_dict = {}
@@ -154,7 +156,7 @@ def convert_nc(output_type, input_file, output_file, band, decimals, band_min, b
             valid_pixels = (eval(converted_vpe).astype(int) * -1) + 1
         except:
             log(env["General"]["log"], "No valid pixel expression for {}".format(band), indent=4)
-            valid_pixels = np.zeros_like(values.flatten())
+            valid_pixels = np.zeros_like(values_flat)
 
     valid_pixels[values_flat < band_min] = 1
     valid_pixels[values_flat > band_max] = 1
@@ -164,15 +166,15 @@ def convert_nc(output_type, input_file, output_file, band, decimals, band_min, b
         log(env["General"]["log"], "Outputting JSON file {}".format(output_file), indent=4)
         out_dict = {band: values_flat}
         df = pd.DataFrame.from_dict(out_dict)
-        if len(lat.shape) > 1:
-            df["lon"] = lon.flatten()
-            df["lat"] = lat.flatten()
-            xmin, ymin, xmax, ymax = [np.nanmin(lon), np.nanmin(lat), np.nanmax(lon), np.nanmax(lat)]
+        if len(y.shape) > 1:
+            df["lon"] = x.flatten()
+            df["y"] = y.flatten()
+            xmin, ymin, xmax, ymax = [np.nanmin(x), np.nanmin(y), np.nanmax(x), np.nanmax(y)]
             lonres, latres = (xmax - xmin) / float(values.shape[0]), (ymax - ymin) / float(values.shape[1])
         else:
-            df["lon"] = np.repeat(lon[np.newaxis, :], len(lat), axis=0).flatten()
-            df["lat"] = np.repeat(lat[:, np.newaxis], len(lon), axis=1).flatten()
-            lonres, latres = float(round(abs(lon[1] - lon[0]), 12)), float(round(abs(lat[1] - lat[0]), 12))
+            df["lon"] = np.repeat(x[np.newaxis, :], len(y), axis=0).flatten()
+            df["lat"] = np.repeat(y[:, np.newaxis], len(x), axis=1).flatten()
+            lonres, latres = float(round(abs(x[1] - x[0]), 12)), float(round(abs(y[1] - y[0]), 12))
         df["valid_pixels"] = valid_pixels
         df.dropna(subset=[band], inplace=True)
         df = df.astype(float).round(decimals)
@@ -181,17 +183,18 @@ def convert_nc(output_type, input_file, output_file, band, decimals, band_min, b
             dump({'lonres': lonres, 'latres': latres, 'lon': list(df["lon"]), 'lat': list(df["lat"]),
                   'v': list(df[band]), 'vp': list(df["valid_pixels"]), 'vpe': valid_pixel_expression,
                   'satellite': satellite, 'datetime': date}, f, separators=(',', ':'))
+
     elif output_type == "geotiff":
         log(env["General"]["log"], "Outputting GEOTIFF file {}".format(output_file), indent=4)
         temp_file = os.path.join(os.path.dirname(output_file), "temp_" + os.path.basename(output_file))
-        if len(lat.shape) > 1:
+        if len(y.shape) > 1:
             image_size = values.shape
             nx = image_size[0]
             ny = image_size[1]
         else:
-            ny = len(lon)
-            nx = len(lat)
-        xmin, ymin, xmax, ymax = [np.nanmin(lon), np.nanmin(lat), np.nanmax(lon), np.nanmax(lat)]
+            ny = len(x)
+            nx = len(y)
+        xmin, ymin, xmax, ymax = [np.nanmin(x), np.nanmin(y), np.nanmax(x), np.nanmax(y)]
         xres = (xmax - xmin) / float(ny)
         yres = (ymax - ymin) / float(nx)
         geotransform = (xmin, xres, 0, ymax, 0, -yres)
@@ -199,14 +202,16 @@ def convert_nc(output_type, input_file, output_file, band, decimals, band_min, b
         dst_ds = gdal.GetDriverByName('GTiff').Create(temp_file, ny, nx, 2, gdal.GDT_Float32)
         dst_ds.SetGeoTransform(geotransform)
         srs = osr.SpatialReference()
-        srs.ImportFromEPSG(4326)
+        srs.ImportFromEPSG(projection)
         dst_ds.SetProjection(srs.ExportToWkt())
         dst_ds.GetRasterBand(1).WriteArray(values)
         dst_ds.GetRasterBand(2).WriteArray(valid_pixels.reshape(values.shape))
-        dst_ds.GetRasterBand(1).SetNoDataValue(np.nan)
-        dst_ds.GetRasterBand(2).SetNoDataValue(np.nan)
         dst_ds.FlushCache()
         dst_ds = None
+
+        if projection != 4326:
+            dsReproj = gdal.Warp(temp_file, temp_file, dstSRS="EPSG:4326", dstNodata=np.nan)
+            dsReproj = None
 
         translate_options = gdal.TranslateOptions(gdal.ParseCommandLine(
             '-co TILED=YES -co COPY_SRC_OVERVIEWS=YES -co COMPRESS=DEFLATE'))
@@ -218,8 +223,13 @@ def convert_nc(output_type, input_file, output_file, band, decimals, band_min, b
         ds = gdal.Open(temp_file)
         gdal.Translate(output_file, ds, options=translate_options)
         ds = None
-        os.unlink(temp_file)
-        os.unlink(temp_file + ".ovr")
+
+        if os.path.isfile(temp_file):
+            os.unlink(temp_file)
+        if os.path.isfile(temp_file + ".ovr"):
+            os.unlink(temp_file + ".ovr")
+        if os.path.isfile(output_file + ".aux.xml"):
+            os.unlink(output_file + ".aux.xml")
 
 
 def parse_bands(bands):
